setup/README.md: Add YAML front matter for Jekyll
The bc repository renders content with Jekyll [1], so we need to add
YAML front matter if we want this file to be processed [2].  This
change is unique to the bc-version of this file; it's not in the
upstream installation-test branch.  That shouldn't be a problem
though, because it's unlikely to conflict with further upstream
development.

[1]: See the root README.md in this repository.
[2]: http://jekyllrb.com/docs/frontmatter/

Reported-by: Aron Ahmadia <aron@ahmadia.net>